refactor: unify config and ruleset naming
