+++ src/rulesets/loader.ts
@@ -1,14 +1,7 @@
 import { join } from '@stoplight/path';
 import * as fs from 'fs';
-import { ILintConfig, OutputFormat } from '../types/config';
 
 const DEFAULT_RULESET_FILE = /^\.?spectral\.(?:ya?ml|json)$/;
-
-export const createEmptyConfig = (): ILintConfig => ({
-  encoding: 'utf8',
-  format: OutputFormat.STYLISH,
-  verbose: false,
-});
 
 export const getDefaultRulesetFile = (directory: string): Promise<string | null> => {
   return new Promise(resolve => {
